backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiomysql>=0.2.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
redis>=5.0.1
httpx>=0.26.0
pytest>=8.0.0
pytest>=8.0.0
email-validator>=2.1.0
bcrypt==3.2.2
